update authorisation - jwt

+++ backend/resources/blog.py
@@ -2,8 +2,10 @@ from datetime import datetime
 from flask import Flask, request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from database import db
 from schemas import BlogSchema, BlogsSchema, BlogEditSchema, BlogDeleteSchema
+
 
 blp = Blueprint("Blogs", __name__, description="Operations on Blogs")
 
@@ -26,25 +28,38 @@ class GetBlogs(MethodView):
 
 @blp.route("/blog")
 class Blogs(MethodView):
+    @jwt_required()
     @blp.response(201, BlogSchema)
     @blp.arguments(BlogSchema)
     def post(self, blog_data):
+        current_user_id = get_jwt_identity()
+        if blog_data["author_id"] != current_user_id:
+            abort(403, "You are not authorized to add blog on this account")
         blog_data["time"] = str(datetime.today())
         response = db.add_blog(blog_data)
         if "error" in response:
             abort(401, response["error"])
         return response
 
+    @jwt_required()
     @blp.response(200, BlogEditSchema)
     @blp.arguments(BlogEditSchema)
     def put(self, blog_data):
+        current_user_id = get_jwt_identity()
+        if blog_data["author_id"] != current_user_id:
+            abort(403, "You are not authorized to edit this blog")
+
         response = db.edit_blog(blog_data)
         if "error" in response:
             abort(401, response["error"])
         return response
 
+    @jwt_required()
     @blp.arguments(BlogDeleteSchema)
     def delete(self, blog_data):
+        current_user_id = get_jwt_identity()
+        if blog_data["author_id"] != current_user_id:
+            abort(403, "You are not authorized to delete this blog")
         response = db.delete_blog(blog_data)
         if "error" in response:
             abort(403, response["error"])
